Check skin/sprite images for correctness
