--- src/engine/shared/dilate.h
@@ -1,6 +1,0 @@
-#ifndef ENGINE_SHARED_DILATE_H
-#define ENGINE_SHARED_DILATE_H
-
-void DilateImage(unsigned char *pImageBuff, int w, int h, int BPP);
-
-#endif
